refactor: Inventory, Promotions 클래스의 내부 메서드를 Parser 클래스로 분리
파싱 관련 메서드 클래스 분리
- parseProductInfo() 메서드
- parsePromotionInfo() 메서드와 toLocalDateTime() 메서드
